requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0
clarifai-grpc==9.9.0
Pillow==10.1.0 
gunicorn
flask
werkzeug==2.2.3
itsdangerous==2.1.2
redis==5.0.1